getCode→getCodeQueryValue・UIAlertControllerのタイトル変更・QiitaModelData分散・OAuthViewControllerDelegate 継承先削除
