+++ Qcli/Extensions/URL+GetCode.swift
@@ -8,12 +8,16 @@
 import Foundation
 
 extension URL {
-    func getCode() -> String? {
+    //リダイレクト先のURLからcodeクエリの値を取得する
+    func getCodeQueryValue() -> String? {
         let codeKey = "code"
         guard self.scheme == AuthorizeKey.scheme.rawValue else { return nil }
         guard let urlComponent = URLComponents(url: self, resolvingAgainstBaseURL: true),
               let queryItems = urlComponent.queryItems,
-              let code = queryItems.filter({ $0.name == codeKey }).compactMap({ $0.value }).first else { return nil }
+              let code = queryItems
+                .filter({ $0.name == codeKey })
+                .compactMap({ $0.value })
+                .first else { return nil }
         return code
     }
 }
